work in progress; completed separation of terminal pane into a TerminalPane, TerminalSessionsPanel, and TerminalSession
